requirements.txt
# Core dependencies
python-decouple
requests
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-redis==6.0.0
django_celery_results==2.6.0
psycopg2-binary==2.9.11
redis==6.4.0
drf-yasg
django-widget-tweaks

# Testing
pytest==8.4.2
pytest-django==4.11.1
factory_boy==3.3.3
Faker==37.11.0
PyYAML==6.0.3

# Code quality
flake8
black
isort
pre-commit
# Security tools (optional)
bandit
safety